// src/app/dashboard/info/interfaces/info.interface.ts
export interface Info {
    id: number;
    companyName: string;
    bossName: string;
    logo: string;
    isActive: boolean;
    sanitaryLicense: string;
    address_street: string;
    address_cp: string;
    address_state: string;
    address_municipality: string;
    address_colony: string;
    sanitarySignature: string;
}